[Auctions] Adds reset button to refine view controller.
