+++ Artsy_Tests/View_Controller_Tests/AuctionRefineViewControllerTests.swift
@@ -7,15 +7,26 @@ import Artsy
 
 class AuctionRefineViewControllerSpec: QuickSpec {
     override func spec() {
+        let defaultSettings = AuctionRefineSettings(ordering: .LotNumber)
+        let differentSettings = AuctionRefineSettings(ordering: .ArtistAlphabetical)
+
         it("looks good by default") {
-            let subject = AuctionRefineViewController()
+            let subject = AuctionRefineViewController(defaultSettings: defaultSettings, initialSettings: defaultSettings)
 
             expect(subject).to( haveValidSnapshot() )
         }
 
         it("looks good when configured with options") {
-            let subject = AuctionRefineViewController()
-            subject.initialSettings = AuctionRefineSettings(ordering: AuctionOrderingSwitchValue.ArtistAlphabetical)
+            let subject = AuctionRefineViewController(defaultSettings: defaultSettings, initialSettings: differentSettings)
+
+            expect(subject).to( haveValidSnapshot() )
+        }
+
+        it("looks good when configured with options and changed options") {
+            let subject = AuctionRefineViewController(defaultSettings: defaultSettings, initialSettings: defaultSettings)
+            
+            subject.loadViewProgrammatically()
+            subject.currentSettings = differentSettings
 
             expect(subject).to( haveValidSnapshot() )
         }
